[docs] Add warning about unsafe-change-config
Adding an explicit warning about unsafe-change-config being unsafe to
avoid users skimming through the docs and deciding this is the command
they need.

Also pointed out that this shouldn't be done if the missing replicas are
a consequence of downed tablet servers and there's a chance of
recovering the tablet servers themselves.

Change-Id: I266c050c237a5c2fa286208a154c2a80f906769c
Reviewed-on: http://gerrit.cloudera.org:8080/9906
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>