build-and-test.sh: unbreak Python 2.6 based builds
The first beta of pip 10 was released several days ago, and it officially
drops support for Python 2.6, going so far as to use syntax that isn't
recognized by Python 2.6 interpreters. Let's address this by pinning to a
specific version of pip when upgrading (which, frankly, we should have done
a long time ago).

Note: pip supports environmental markers in version specifiers which we
could use to narrow this pinning such that it only takes effect if we're on
Python 2.6. However, the version of pip typically found in older distros is
old enough that it doesn't support environmental markers, so we'd need to do
two pip upgrades: the first to land on a version of pip that supports
markers, and the second to use the markers to conditionally upgrade to the
latest pip. This didn't seem worth the hassle.

Change-Id: I1afb5ba8169d193f502d3733fd3f826e9cb4ce6e
Reviewed-on: http://gerrit.cloudera.org:8080/9936
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>