[docs] Add lpeers output to ntp monitoring
The ntp monitoring had an example of opeers output and had a tip
explaining some versions have lpeers and some others opeers. These two
are in fact two separate, although similar commands, and as far as I
know both of them are available in all modern versions.

Added an lpeers output, but had to change the lpeers output as well to
keep things consistent. I also changed the tip to reflect this.

Change-Id: I5009a330cfd3e81496e95b3e3cf1fb2b7627b085
Reviewed-on: http://gerrit.cloudera.org:8080/10041
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>