[Java] Mark the Yetus annotations dependency optional
Given the Yetus dependency is mainly used for documentation
generation and is not used at runtime, we can mark the
annotation as optional for users consuming Kudu.
This is similar to the jsr305 dependency which we mark
as optional.

Additionally we don’t need to include Yetus in our shaded
jars.

Change-Id: I6daf4f21c6cdde08145f5467e84d68dbe40cec16
Reviewed-on: http://gerrit.cloudera.org:8080/10095
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins