[Java] Use Yetus annotations in place of @VisibleForTesting
We use yetus for marking our intended InterfaceAudience.
This patch replaces Guava’s @VisibleForTesting with
@InterfaceAudience.LimitedPrivate("Test”).

kudu-hive is an exception to the rule because it can not
have any dependencies that are not provided by the
hive-metastore. This limitation is noted in the kudu-hive
build files.

I also elimated an uneccesary usage of ImmutableList
and Lists while making this change. Scala has
straightforward syntax for making lists without Guava.

The primary reason for this change is to use
Yetus as the single annotation library for defining
InterfaceAudience, but this has the added benefit
of eliminating any Guava use in kudu-spark,
kudu-spark-tools which didn’t
have Gauva marked as a dependency.

Change-Id: Ic0d8845e8cbbf5ea81345c46c5487121d1098f91
Reviewed-on: http://gerrit.cloudera.org:8080/10077
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Dan Burkert <danburkert@apache.org>