Modify fix_includes.py for Kudu usage, add a wrapper
This makes a few small changes to the upstream fix_includes.py script:

- We name our tests 'foo-test.cc' instead of 'foo_test.cc'. Expands
  the regex appropriately.

- Add support for 'IWYU pragma: keep' to avoid removing duplicate
  headers. Without this, it was attempting to remove the duplicate
  include in bitshuffle_arch_wrapper.cc which includes the bitshuffle
  header twice with different #defines

- Adds a --blank_line_between_c_and_cxx_includes option which preserves
  the style that we use.

- Adds support for differentiating between thirdparty <foo> includes
  versus system <foo> includes by checking for file existence in a set
  of directories passed on the command line. Google apparently uses
  quoted includes like "protobuf/foo.h" for this purpose whereas we use
  <>.

This also adds a wrapper script build-support/iwyu.py which handles
invoking IWYU and also integrating it with the modified fix_includes.py.

This takes the place of the old 'iwyu.sh' and associated awk script. It
supports automatically piping the IWYU results back through the include
fixer and either producing a diff of errors (default output) or fixing
them in-situ.

Change-Id: I3c286271a39a0d825fb11e5610d8eb7e5b0729b9
Reviewed-on: http://gerrit.cloudera.org:8080/10106
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>