iwyu: pass source root into fix_includes
fix_includes tries to identify the "main compilation unit header" for
each file using some heuristics. For example, "foo-test.cc" is
associated with "foo.h".

This was previously broken due to the fact that the tool runs from the
build/ directory instead of the source directory, so it saw the CC file
as "../../src/kudu/.../foo.cc" while it saw the header file as
"kudu/.../foo.h".

This adds a --source_root parameter to fix_includes so that the source
files can be relativized to this path prior to fixing.

Change-Id: Iac9f0b47b9a3c8180e178ed59863d33840434433
Reviewed-on: http://gerrit.cloudera.org:8080/10159
Tested-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>