iwyu: add proper dependency on thrift generated code, check for iwyu errors
IWYU was giving incorrect results on the HMS modules when running on
Jenkins. This was due to the 'iwyu' CMake target not depending on the
thrift code generation. IWYU was continuing to run despite having
'missing include' errors.

This patch adds the appropriate CMake dependency and also makes our
IWYU wrapper check for such errors in the IWYU output.

I also changed the IWYU wrapper to explicitly build the IWYU
dependencies before running so that even if a user invokes it directly,
they won't hit this issue.

Change-Id: Ie44f25ebc160c587eb11b4b4730095ee1e6c4fde
Reviewed-on: http://gerrit.cloudera.org:8080/10161
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>