KUDU-2191: Metadata Upgrade Tool
This commit introduces an upgrade tool to allows backward compatibility.
It provides support to upgrade the metadata format of existing Impala
managed/external tables in HMS entries, as well as rename the existing
tables in Kudu to adapt to the new naming rules. To run this tool, it
requires to turn off the HMS integration feature. It adds test case
using external mini cluster to ensure the upgrade tool works as
expected.

Change-Id: I1c8488e1c7b04b8bedc31c83b1496b53bae49cb3
Reviewed-on: http://gerrit.cloudera.org:8080/10075
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins