[Java] Remove shaded parquet and spark-avro dependencies
Currently the kudu-client-tools project shades parquet and
the kudu-spark-tools project shades spark-avro. According
to commit history this is done for classpath convenience when
the parquet and avro import/export options are used.

However, our current shading configuration in Maven is not
pulling in transitive dependencies but instead only including
the direct classes from those jars. If it works at runtime it's just
luck that the rest of the classes are there or unused.

The Gradle build is currently including all of the transtive
dependencies, but that results in a jar that is very large
including things like the Scala library, Jackson, Snappy, etc.

Instead of packaging/shading such large dependencies,
this patch changes the dependencies to provided scope which
accurately representing the fact that the libarries are expected
to be on the classpath at runtime and document’s the details.

Change-Id: Iccf46be9eebb91e900a9ebb5f99b6510165956e7
Reviewed-on: http://gerrit.cloudera.org:8080/10147
Tested-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>