KUDU-2191 (11/n): remove fallback logic from HmsCatalog
HmsCatalog was originally written with fallback logic that would
liberally correct HMS entries when possible. In particular, altering a
non-existent table entry would result in the HMS Catalog creating a new
table entry. These liberal corrections allowed the HMS catalog to
'upgrade' legacy Kudu tables to have HMS entries automatically upon any
kind of alteration, but in certain race conditions[1] it could lead to a
single Kudu table having multiple HMS table entries with different
names. This will be problematic down the line for other integrations
like Sentry, which rely on HMS metadata being canonical.

This commit removes this logic, which simplifies HmsCatalog, and removes
the need for a bunch of the tests which covered this fallback logic. As
a result, the metadata upgrade tool is going to need a special path
which allows it to alter Kudu tables without modifying the HMS. That
will be included in a follow up commit.

[1]: consider the following operation sequence:

      - in Hive: ALTER TABLE a RENAME TO b
      - in Kudu: ALTER TABLE a RENAME TO c

  When handling the Kudu alter table, the 'a' table exists in the Kudu
  catalog, but not in the HMS. As a result the HmsCatalog creates a new
  table entry with name 'c'. At this point the HMS contains two table
  entries ('b' and 'c'), both representing the same Kudu table. This
  will cause problems for Sentry, for dropping the table, etc.

Change-Id: I0753b1868b406c32e183896caa0427eedebdd8f7
Reviewed-on: http://gerrit.cloudera.org:8080/10380
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>