KUDU-2414. Scanner should not retry a SCANNER_EXPIRED error without reopening
This fixes yet another tight retry loop seen while testing Impala. A scanner
expired on the server side, and the client got the expiration error. However,
the current code sees that as retryable, but should know that it's the kind
of retry that must re-open the scanner.

Without the fix, the new unit test fails with:

../../src/kudu/client/client-test.cc:1504: Failure
      Expected: "Not found: Scanner not found"
To be equal to: s.ToString()
      Which is: "Timed out: Scan RPC to 127.110.140.1:33568 timed out after 0.000s (SENT): Not found: Scanner not found"
../../src/kudu/client/client-test.cc:1508: Failure
      Expected: 1
To be equal to: scan_rpcs->TotalCount()
      Which is: 104127

Change-Id: I5e09849d5b749db958c3a3df89fd8de9947ab991
Reviewed-on: http://gerrit.cloudera.org:8080/2711
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>