[Java] Check in the Gradle wrapper properties
We were downloading the properties file, but that can result in
using an outdated Gradle wrapper distribution.

This changes .gitignore so the generated
gradle-wrapper.properties file is checked in anytime
we change the gradle verison and updates wrapper.gradle
to add a license to the header.

I also added the silent (-s) flag to the curl command
when downloading the jar to remove some noise from
the log.

Change-Id: I3a772d5e093be29fcbabcbd8324bd6434fb33beb
Reviewed-on: http://gerrit.cloudera.org:8080/10215
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>