KUDU-2460. build: improve parse_test_failure script
This updates the script to properly catch UBSAN and LSAN errors. It
additionally does a bunch of refactoring and cleanup and adds a couple
simple unit tests for the new functionality.

The main reason for the refactoring is that I hope to repurpose this
script to be run by a Python port of run-test.py in order to look for
things like TSAN/ASAN/UBSAN failures and make sure they end up in an
.xml file so that Jenkins surfaces them more readily.

Change-Id: Idbc5ce7f7aa522c2c581afe33d4e40719ed7474a
Reviewed-on: http://gerrit.cloudera.org:8080/10224
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@apache.org>