[tests] fixed flake in consensus_peer_health_status
Fixed flake in the TestPeerHealthStatusTransitions scenario of the
ConsensusPeerHealthStatusITest test.  Prior to the fix, the flakiness
happened when the target tablet server was shutdown during an on-going
tablet copy, where the tablet copy was initiated by AddServer() call
while preparing the mini-cluster for the peer health sequence of
"HEALTHY -> UNKNOWN -> FAILED -> FAILED_UNRECOVERABLE".
In that situation, the source tablet server had corresponding WAL
segments anchored, so they could not be GCed.  As a result, the tablet
replica would not get FAILED_UNRECOVERABLE health status in 30 seconds
because --tablet_copy_idle_timeout_sec is set to 600 seconds by default.

Test results using dist-test, before and after the fix (ASAN),
before:
  http://dist-test.cloudera.org/job?job_id=aserbin.1525111507.120645

after:
  http://dist-test.cloudera.org/job?job_id=aserbin.1525099526.63998

Change-Id: I8eb640604e98361029aa3342ffa3050e922b6629
Reviewed-on: http://gerrit.cloudera.org:8080/10237
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>