Add validation and docs on setting block_cache_capacity_mb too high
Raising --block_cache_capacity_mb can improve Kudu's performance, but
raising it too high (above the memory pressure threshold, which is
--memory_pressure_percentage of the hard memory limit) causes
constant flushing even if write throughput is low. This patch adds
validation of the capacity using a group flag validator, and does
not permit Kudu to start if the block cache size can can cause
Kudu to get stuck in aggressive flushing mode. It warns if the
capacity is large enough (> 50%) of the pressure threshold.

I chose 50% as the max recommended block cache size as a
percentage of the memory pressure threshold somewhat arbitrarily.
It seems safe, but also leaves room for allocating a large
block cache.

There's no automated tests but I did verify the expected messages
appeared in the logs and the server fails to start when
--block_cache_capacity_mb is too high.

Change-Id: Ia99fa95c578235beb3a1c72a33a22b4d8ff4800c
Reviewed-on: http://gerrit.cloudera.org:8080/10266
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins