[jepsen] make kudu-master-see-tservers more robust
Updated kudu-master-see-tservers function to be less fragile.
The prior implementation depended on the format of the output of the
'kudu cluster ksck' sub-command.  The updated implementation uses
'kudu tserver list' and relies only on the number of output lines
by the latter sub-command.

Also, dropped redundant 'node' parameter of the kudu-cli function.

Change-Id: I8b9ff0b7c145e5235a94de27bbedea1fe36f5e60
Reviewed-on: http://gerrit.cloudera.org:8080/10290
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>