KUDU-2191: HMS inconsistent metadata fix tool
This commit introduces a CLI tool that fixes inconsistent metadata for
Hive Metastore integration. It corrects the table name based on the
users input. And it fixes the master addresses property, column name
and type based on the values in Kudu. Test case is added using external
mini cluster to ensure the tool works as expected.

Change-Id: I63c694b5d9877cfbd218139f2e9ecf05fd69c997
Reviewed-on: http://gerrit.cloudera.org:8080/10303
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>