[jepsen] limit block cache capacity to 128MB
Changelist 297b72b introduced consistency check for the relative ratio
of the block cache capacity and the memory pressure threshold.  Since
the default capacity of the block cache is 512MB, that lead to inability
to start tablet servers and masters on machines with relatively small
amount of total memory.

This changelist set the block cache capacity limit to 128MB for
Kudu Jepsen tests.  As of now, Kudu Jepsen tests do not require
a huge block cache capacity.

Change-Id: Iebc6c9b3c616270dd233be4474cb0b29068efda9
Reviewed-on: http://gerrit.cloudera.org:8080/10308
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>