[rebalancing] add a rebalancing algorithm
This changelist introduces the high-level rebalancing algorithm
and corresponding unit tests. It contains both an abstraction for
rebalancing algorithms in general and an implementation of a
greedy algorithm that rebalances a cluster by trying to equalize
the number of replicas per tablet server for each table and for the
cluster as a whole.

The algorithm just identifies what moves should be done given an
arrangement of a cluster. The 'engine' that performs moves
will be added in a subsequent changelist.

Change-Id: I5a8050ee79117a2ae2f6f88740ed25656946cfb4
Reviewed-on: http://gerrit.cloudera.org:8080/10336
Reviewed-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
Tested-by: Kudu Jenkins