[docs] Add a bit of information about tombstoned replicas
This adds a quick paragraph to the troubleshooting docs about tombstoned
replicas. I've seen quite a few users be alarmed by tombstones and
believe they need to be dealt with somehow, possibly by deletion. Of
course, this is dangerous, and having docs to point people towards will
be nice, and may prevent some user consternation.

Change-Id: I876aeb1ffeb7251fe36fe6022b722ed88997cdbe
Reviewed-on: http://gerrit.cloudera.org:8080/10372
Tested-by: Will Berkeley <wdberkeley@gmail.com>
Reviewed-by: Mike Percy <mpercy@apache.org>