[tools] rebalancer in the kudu CLI tool
Introduced rebalancing tool as a part of the Kudu CLI tool.  To run
the rebalancer against a Kudu cluster:

  sudo -u kudu kudu cluster rebalance <master_addresses>

This changelist also contains unit tests to cover converting KsckResults
into ClusterBalanceInfo and integration tests to verify
the functionality of the rebalancer as a part of the Kudu CLI tool.

The code was tested against a 12-node cluster running the recent
Kudu 1.7 (3-4-3 replica management scheme) with about 12TB of data total
and a cluster running Kudu 1.4 (3-2-3 replica management scheme).

This tool works against Kudu clusters of version 1.4 and above.
It does not work against Kudu version 1.3 and earlier.

Change-Id: I269ea1dcb0b528ad9f03308bac6b8769e2141238
Reviewed-on: http://gerrit.cloudera.org:8080/10399
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>