KUDU-2427: fix glog symbolization in certain ASLR environments
On Ubuntu 18.04, glog is unable to symbolize the stack frames belonging to
a unit test binary itself, leaving them as "(unknown)". This breaks various
unit tests that check the contents of those frames.

There hasn't been a new glog release in some time, but this upstream patch
fixes the problem.

Change-Id: Iba6f715123ddf41ba67b93860caf7041b8e137c4
Reviewed-on: http://gerrit.cloudera.org:8080/10431
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>